requirements.txt
requests==2.31.0
boto3==1.34.0
botocore==1.34.0
pandas>=2.2.2
pyarrow>=15.0.0
python-dotenv==1.0.0
pytest>=8.0.0